feat: Use extras to install specific dependencies
The only missing piece is MQTT, it is way to intriqued everywhere.
So for now, the `paho-mqtt` dependency will be installed by default.

At some point, it will be cool to be able to get rid of it and install it via:

```shell
$ python -m pip install 'mppsolar[mqtt]'
```

Also, I added very primary tests to see if all protocols and outputs can be
imported even when a depedency is missing. A proper message will be printed
in that case (except for `esp32io.py` and the `machine` import, I guss this is very specific?).